+++ src/components/recreation/RecreationReview.jsx
@@ -4,7 +4,7 @@ import yellowStar from '../../assets/recreation/yellowStar.svg';
 import greyStar from '../../assets/recreation/greyStar.svg';
 import { ReactComponent as GoodIcon } from '../../assets/recreation/good.svg';
 import { ReactComponent as BadIcon } from '../../assets/recreation/bad.svg';
-import { getRelativeTimeText } from '../../utils/time';
+import { getRelativeTimeText } from '../../utils/timeUtils';
 import { privateAPI } from '../../apis/user';
 import useLoginModalStore from '../../stores/loginModalStore';
 import useLoginStore from '../../stores/loginStore';
